## Description

Moving an interface from one namespace to another is time consuming (especially in CANAL when trying to create multiple pods) and requires the interface to be set back to UP again.  We can improve the current implementation by creating the host veth directly in the host namespace.

We measured the time it takes in calico cni plugin to set up the networking stack, more precisely the time it takes to execute DoNetworking. We ran benchmarks measuring the time spent when creating 16 pods simultaneously on a cluster with 1 control-plane node and 1 worker node. 

<details>
<summary>System specs</summary>
Red Hat Enterprise Linux 9.1
Kernel: 6.3.0-rc3
</details>

Below are the benchmarks for various Calico setups:
<details>
<summary>CALICO IPV4</summary>

```
before:
containerd[5219]: 2023-05-26 07:58:13.018 [INFO][46450] dataplane_linux.go 319: Finished DoNetworking. ContainerID="276c70d6b1f9aaa0b012399b5c99c97ec5abf6f47ef9b7655855ed04ad515e90" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-5" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--5-eth0" after=15.357635ms
containerd[5219]: 2023-05-26 07:58:13.043 [INFO][46448] dataplane_linux.go 319: Finished DoNetworking. ContainerID="6b3c632b3562577523a5acf4389f930be989aef01e1238a3ef0b29818f88a9fd" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-7" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--7-eth0" after=22.190691ms
containerd[5219]: 2023-05-26 07:58:13.071 [INFO][46442] dataplane_linux.go 319: Finished DoNetworking. ContainerID="718ea8a154349f0339c349f3e8ae6cded7bd8deaf171d864ffb8e50ce87434d9" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-2" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--2-eth0" after=28.392989ms
containerd[5219]: 2023-05-26 07:58:13.088 [INFO][46468] dataplane_linux.go 319: Finished DoNetworking. ContainerID="9abbe9758a181cc8f78cd9fe578e07506921fd1c8cbba7ee6f7b624ea3319726" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-4" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--4-eth0" after=28.349732ms
containerd[5219]: 2023-05-26 07:58:13.101 [INFO][46456] dataplane_linux.go 319: Finished DoNetworking. ContainerID="735bc443d42817f0dbb2cc09354913128f84ced50e5418464eae50187d01bd45" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-13" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--13-eth0" after=23.831761ms
containerd[5219]: 2023-05-26 07:58:13.152 [INFO][46547] dataplane_linux.go 319: Finished DoNetworking. ContainerID="23c690d255b131813d73eb7a5ddaa70f443a2a9b487610e3b878de8f3bce8d80" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-11" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--11-eth0" after=56.314785ms
containerd[5219]: 2023-05-26 07:58:13.165 [INFO][46542] dataplane_linux.go 319: Finished DoNetworking. ContainerID="6b77a64747716e78037105a116289993ce5d63451f9e052780726cb8682eb442" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-3" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--3-eth0" after=27.449413ms
containerd[5219]: 2023-05-26 07:58:13.165 [INFO][46555] dataplane_linux.go 319: Finished DoNetworking. ContainerID="f17776dda765527b52fa04d2710c3bd50ff7ec40aa26807d72b5720f1e761113" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-15" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--15-eth0" after=49.489791ms
containerd[5219]: 2023-05-26 07:58:13.178 [INFO][46544] dataplane_linux.go 319: Finished DoNetworking. ContainerID="7c6cb359601d07ab317015187f11eb41b716a031db0b423635a226d794e71fa8" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-14" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--14-eth0" after=14.242354ms
containerd[5219]: 2023-05-26 07:58:13.200 [INFO][46522] dataplane_linux.go 319: Finished DoNetworking. ContainerID="96a74315693e51f4e073daa889a89cc96143a45f3c9f88e239f515ab70b331ee" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-8" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--8-eth0" after=13.880612ms
containerd[5219]: 2023-05-26 07:58:13.225 [INFO][46609] dataplane_linux.go 319: Finished DoNetworking. ContainerID="7fb087f7320822694b9d5ad82650c50e82a6084ffe628c03a0f35dcac477d547" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-0" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--0-eth0" after=19.716989ms
containerd[5219]: 2023-05-26 07:58:13.246 [INFO][46585] dataplane_linux.go 319: Finished DoNetworking. ContainerID="9907596d244db3f61c92fdaad26fe8701ad7e73a4372abfdd8dc0dc33c890586" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-1" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--1-eth0" after=18.306579ms
containerd[5219]: 2023-05-26 07:58:13.261 [INFO][46684] dataplane_linux.go 319: Finished DoNetworking. ContainerID="2370e8ebcda79939e88e9a00046b962192139ed2f74636e49cde2e5df97e04f2" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-10" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--10-eth0" after=11.811679ms
containerd[5219]: 2023-05-26 07:58:13.288 [INFO][46641] dataplane_linux.go 319: Finished DoNetworking. ContainerID="de053a76856fd216e2d23e6f3953ffe587fc78689733afe8fcc603827f13b48a" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-12" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--12-eth0" after=16.934841ms
containerd[5219]: 2023-05-26 07:58:13.300 [INFO][46686] dataplane_linux.go 319: Finished DoNetworking. ContainerID="e4b6baf78b1531cc14d6f62184aaa5e21cf5364fc045412fe4a996d6d92789df" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-6" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--6-eth0" after=10.119797ms
containerd[5219]: 2023-05-26 07:58:13.332 [INFO][46679] dataplane_linux.go 319: Finished DoNetworking. ContainerID="251b511582a5392aeffe51e32027c6b0bb57241fec1674a0ba5bab24c8d805a6" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-9" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--9-eth0" after=19.631156ms

after:
containerd[5219]: 2023-05-26 08:00:34.329 [INFO][53745] dataplane_linux.go 329: Finished DoNetworking. ContainerID="5e2dd6e3715a5bad939e77c1483b8b6ac1dac3b4d9898e03357438992ecd017e" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-0" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--0-eth0" after=3.337002ms
containerd[5219]: 2023-05-26 08:00:34.349 [INFO][53780] dataplane_linux.go 329: Finished DoNetworking. ContainerID="bf9ec3565e7dbcff37127f15562538708672c49e22a80a4e6e56da78bc7c7ef2" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-13" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--13-eth0" after=1.477435ms
containerd[5219]: 2023-05-26 08:00:34.369 [INFO][53723] dataplane_linux.go 329: Finished DoNetworking. ContainerID="f3305dd83031e0d5b40996af603a9f946ba2a0a04ecbb505e378406b8d0fb2bc" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-1" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--1-eth0" after=2.212617ms
containerd[5219]: 2023-05-26 08:00:34.387 [INFO][53766] dataplane_linux.go 329: Finished DoNetworking. ContainerID="2da2353bb55767dbbe0988cbb40ff1362c68d1045c02d3152c1d1db6d1ddb7be" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-5" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--5-eth0" after=1.002398ms
containerd[5219]: 2023-05-26 08:00:34.406 [INFO][53792] dataplane_linux.go 329: Finished DoNetworking. ContainerID="0c6f2af44a40e47af4df235d65721df42a9c8cc868a6562e56e8b6f84d163660" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-2" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--2-eth0" after=1.103331ms
containerd[5219]: 2023-05-26 08:00:34.426 [INFO][53752] dataplane_linux.go 329: Finished DoNetworking. ContainerID="7dd4a90d2a8950781a635c5e1a65f3dd18724d82599b4ab80194bb513ea7bbd4" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-15" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--15-eth0" after=1.151782ms
containerd[5219]: 2023-05-26 08:00:34.447 [INFO][53788] dataplane_linux.go 329: Finished DoNetworking. ContainerID="2606d5f45d4c4ea406fc5e501e7f99c69e227b00066fc4685942964ff579126e" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-14" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--14-eth0" after=1.213658ms
containerd[5219]: 2023-05-26 08:00:34.466 [INFO][53720] dataplane_linux.go 329: Finished DoNetworking. ContainerID="f9215dfc55398852496bc93feca7496f6fa41686d1e9329d86ef535f5d4c70ce" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-10" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--10-eth0" after=1.04141ms
containerd[5219]: 2023-05-26 08:00:34.484 [INFO][53935] dataplane_linux.go 329: Finished DoNetworking. ContainerID="b408c5f2d1269dd75e6f561687ef1119d116396fb966b59c016b1f21a2c6d685" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-12" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--12-eth0" after=1.139266ms
containerd[5219]: 2023-05-26 08:00:34.507 [INFO][53956] dataplane_linux.go 329: Finished DoNetworking. ContainerID="d096353e7e40d4e7e6911d9f381188dba5428a3d1d0d32acfe1889a5f5ea8d93" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-4" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--4-eth0" after=1.285217ms
containerd[5219]: 2023-05-26 08:00:34.534 [INFO][53949] dataplane_linux.go 329: Finished DoNetworking. ContainerID="922ec29998f207b97a900f95bd30a0ad817ccd4f672eabe0c8d1a63e0aeed3a7" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-9" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--9-eth0" after=1.115956ms
containerd[5219]: 2023-05-26 08:00:34.559 [INFO][53918] dataplane_linux.go 329: Finished DoNetworking. ContainerID="cc5b67baee7e21dbf314e4360bef33eab90ba92dd34261845036ffe4d8156065" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-7" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--7-eth0" after=1.05643ms
containerd[5219]: 2023-05-26 08:00:34.582 [INFO][53957] dataplane_linux.go 329: Finished DoNetworking. ContainerID="c9e3bd434f7b42d0d065dbc8f50d13b3b70e5b2f322d5d8e625346c42fc1fe4e" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-6" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--6-eth0" after=1.110293ms
containerd[5219]: 2023-05-26 08:00:34.601 [INFO][53852] dataplane_linux.go 329: Finished DoNetworking. ContainerID="2da63e3965def05b1523d9f8d06452e828d122f88bfe4dab358ff377d0ac72c4" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-8" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--8-eth0" after=906.848µs
containerd[5219]: 2023-05-26 08:00:34.621 [INFO][53871] dataplane_linux.go 329: Finished DoNetworking. ContainerID="b5b891f46cdae179e6a042453f78eb22e67fddff0b940593a3bdc1913db00e56" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-3" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--3-eth0" after=980.648µs
containerd[5219]: 2023-05-26 08:00:34.642 [INFO][53965] dataplane_linux.go 329: Finished DoNetworking. ContainerID="eaf48bfb23e22a57bf12425f019e6de980b98d86c5d470670204f61969a907b6" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-11" WorkloadEndpoint="robot--all--0a65ff80--7adb--4852--9247--5272deaa4bb7--vmss000000-k8s-job--kbench--pod--oid--0--tid--11-eth0" after=1.207534ms
```
</details>

<details>
<summary>CALICO IPV4 + IPV6</summary>

```
before:
containerd[5248]: 2023-05-26 07:30:31.770 [INFO][711285] dataplane_linux.go 319: Finished DoNetworking. ContainerID="74aecf84c455d7934fc39363d15c3622431b5f105a91f8d0309a8cced83cc20e" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-5" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--5-eth0" after=12.629326ms
containerd[5248]: 2023-05-26 07:30:31.837 [INFO][711297] dataplane_linux.go 319: Finished DoNetworking. ContainerID="f028857e80dde0c2b68c894d7c8467309c45d7b3f0fa8f9d99bcdcd4ec818024" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-3" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--3-eth0" after=36.905691ms
containerd[5248]: 2023-05-26 07:30:31.854 [INFO][711312] dataplane_linux.go 319: Finished DoNetworking. ContainerID="a9bae3e7e4120f3b2e987437277728a2d7196790a9586f99312ca61ab1ef1d41" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-13" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--13-eth0" after=18.676218ms
containerd[5248]: 2023-05-26 07:30:31.895 [INFO][711329] dataplane_linux.go 319: Finished DoNetworking. ContainerID="fba246b731812d37576375cab06cfa242b63afc6909a76060bd134c3c6b60f38" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-12" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--12-eth0" after=17.84298ms
containerd[5248]: 2023-05-26 07:30:31.940 [INFO][711325] dataplane_linux.go 319: Finished DoNetworking. ContainerID="cd5f83f64b47047cbb005b4cd7aabeae3e448d37a422000906ab04b6156607ca" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-11" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--11-eth0" after=24.006812ms
containerd[5248]: 2023-05-26 07:30:31.974 [INFO][711338] dataplane_linux.go 319: Finished DoNetworking. ContainerID="0babe843b1d154e2c5a7439e9c0ce83bd3ac402eb554ff581987ea4bf0b21675" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-15" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--15-eth0" after=12.206808ms
containerd[5248]: 2023-05-26 07:30:32.013 [INFO][711373] dataplane_linux.go 319: Finished DoNetworking. ContainerID="2eb4bec12fb07f9d2acb9abb0b331386abcc8980058ba4230ffd1eea2fb9ee12" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-7" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--7-eth0" after=14.885459ms
containerd[5248]: 2023-05-26 07:30:32.057 [INFO][711368] dataplane_linux.go 319: Finished DoNetworking. ContainerID="061a7ab1f9278d797dde41db3504ba86758a102d4b3daaf4108f5e6667ce429b" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-6" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--6-eth0" after=18.818314ms
containerd[5248]: 2023-05-26 07:30:32.092 [INFO][711520] dataplane_linux.go 319: Finished DoNetworking. ContainerID="89622b323e99c78cae452f9cad551333e60099e424ff6d534d23fa68fe87d09c" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-1" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--1-eth0" after=13.376886ms
containerd[5248]: 2023-05-26 07:30:32.132 [INFO][711504] dataplane_linux.go 319: Finished DoNetworking. ContainerID="2e4f01b610b1d287cff45d33130a5e59290ebd7bbae218b017a67215a1675249" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-10" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--10-eth0" after=12.917016ms
containerd[5248]: 2023-05-26 07:30:32.172 [INFO][711540] dataplane_linux.go 319: Finished DoNetworking. ContainerID="5dd1c28e7d26cb05341f01ba66569e8d9a1c62e27f9c72a06a4f9f78e0e047c5" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-8" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--8-eth0" after=12.757143ms
containerd[5248]: 2023-05-26 07:30:32.217 [INFO][711576] dataplane_linux.go 319: Finished DoNetworking. ContainerID="a1ee58ac055345fe423b40fd35ff6e5964f49f46e4cc33f488eb6760b706e853" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-2" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--2-eth0" after=17.519774ms
containerd[5248]: 2023-05-26 07:30:32.278 [INFO][711562] dataplane_linux.go 319: Finished DoNetworking. ContainerID="204fd4455b4aeb14082fe5efb9b15e652f8b98a6e2195109dd692ce7a17541fc" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-9" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--9-eth0" after=20.268953ms
containerd[5248]: 2023-05-26 07:30:32.323 [INFO][711534] dataplane_linux.go 319: Finished DoNetworking. ContainerID="6281c43aef3e10d0b680d30910f89ea6c25a720526729df15dbd72db1d308c2e" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-0" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--0-eth0" after=17.864858ms
containerd[5248]: 2023-05-26 07:30:32.361 [INFO][711583] dataplane_linux.go 319: Finished DoNetworking. ContainerID="da0939b5d2e0c2caee73c5077ceef596c44692d1e7a0d48fbff2011ed03f7265" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-4" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--4-eth0" after=14.855653ms
containerd[5248]: 2023-05-26 07:30:32.400 [INFO][711589] dataplane_linux.go 319: Finished DoNetworking. ContainerID="359551220b23588136dd6eadc96e27e1c347812bff815ec4558b385312cd7c52" Namespace="kbench-pod-namespace" Pod="job-kbench-pod-oid-0-tid-14" WorkloadEndpoint="robot--all--b65806d0--d068--45f4--b8d8--c27894da3aea--vmss000000-k8s-job--kbench--pod--oid--0--tid--14-eth0" after=13.305052ms

after:

…